Refactor: 사용자 관련 클래스를 `domain/user`로 이동
