+++ src/main/java/com/team7/retriever/domain/user/controller/dto/request/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.dto.request;
+package com.team7.retriever.domain.user.controller.dto.request;
 
 public record SignUpRequest(
 	String employeeId,
