+++ src/main/java/com/team7/retriever/domain/user/domain/repository/UserRepository.java
@@ -1,11 +1,11 @@
-package com.team7.retriever.repository;
+package com.team7.retriever.domain.user.domain.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.team7.retriever.entity.User;
+import com.team7.retriever.domain.user.domain.document.User;
 
 @Repository
 public interface UserRepository extends MongoRepository<User, String> {
