--- src/main/java/com/team7/retriever/dto/request/GrantRequest.java
@@ -1,9 +1,0 @@
-package com.team7.retriever.dto.request;
-
-import com.team7.retriever.entity.enums.Role;
-
-public record GrantRequest(
-	String employeeId,
-	Role role
-) {
-}
