+++ src/main/java/com/team7/retriever/global/auth/jwt/provider/JwtTokenProvider.java
@@ -11,7 +11,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.stereotype.Service;
 
-import com.team7.retriever.entity.enums.Role;
+import com.team7.retriever.domain.user.domain.enums.Role;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
